# backend/swagger.yaml
openapi: 3.0.0
info:
   title: صيدليات الشافي API
   description: التوثيق الرسمي لـ REST API الخاصة بتطبيق صيدليات الشافي.
   version: "2.0.0"
servers:
   - url: http://localhost:5000/api
     description: Local development server
   - url: https://localhost:5000/api
     description: Local HTTPS development server
paths:
...
   /products:
      get:
         tags: [Products]
         summary: جلب كل المنتجات
         responses:
            "200":
               description: قائمة المنتجات
               content:
                  application/json:
                     schema:
                        type: array
                        items:
                           $ref: '#/components/schemas/Product'
      post:
         tags: [Products]
         summary: إضافة منتج جديد (للمدراء فقط)
         security:
            - bearerAuth: []
         requestBody:
            required: true
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/ProductInput'
         responses:
            "201":
               description: تم إنشاء المنتج بنجاح
            "403":
               description: غير مصرح لك بالوصول
   /products/{productId}:
      get:
         tags: [Products]
         summary: تفاصيل منتج معين
         parameters:
            - name: productId
              in: path
              required: true
              schema:
                 type: string
         responses:
            "200":
               description: تفاصيل المنتج
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Product'
            "404":
               description: المنتج غير موجود

   # Category Routes
   /categories:
      get:
         tags: [Categories]
         summary: جلب جميع الفئات
         responses:
            "200":
               description: قائمة الفئات
               content:
                  application/json:
                     schema:
                        type: array
                        items:
                           $ref: '#/components/schemas/Category'
      post:
         tags: [Categories]
         summary: إنشاء فئة جديدة
         security:
            - bearerAuth: []
         requestBody:
            required: true
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/CategoryInput'
         responses:
            "201":
               description: تم إنشاء الفئة بنجاح
   /categories/{id}:
      get:
         tags: [Categories]
         summary: جلب فئة معينة
         parameters:
            - name: id
              in: path
              required: true
              schema:
                 type: string
         responses:
            "200":
               description: تفاصيل الفئة
      put:
         tags: [Categories]
         summary: تحديث فئة
         security:
            - bearerAuth: []
         parameters:
            - name: id
              in: path
              required: true
              schema:
                 type: string
         requestBody:
            required: true
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/CategoryInput'
         responses:
            "200":
               description: تم تحديث الفئة بنجاح
      delete:
         tags: [Categories]
         summary: حذف فئة
         security:
            - bearerAuth: []
         parameters:
            - name: id
              in: path
              required: true
              schema:
                 type: string
         responses:
            "200":
               description: تم حذف الفئة بنجاح

   # Cart Routes
...
   /orders:
      get:
         tags: [Orders, Admin]
         summary: جلب جميع الطلبات (للمدراء فقط)
         security:
            - bearerAuth: []
         responses:
            "200":
               description: قائمة جميع الطلبات
            "403":
               description: غير مصرح لك بالوصول
      post:
         tags: [Orders]
         summary: إنشاء طلب جديد
         security:
            - bearerAuth: []
         requestBody:
            required: true
            content:
               application/json:
                  schema:
                     type: object
                     required:
                        - items
                        - shippingAddress
                        - paymentMethod
                     properties:
                        items:
                           type: array
                           items:
                              type: object
                              properties:
                                 productId:
                                    type: string
                                 quantity:
                                    type: integer
                                 price:
                                    type: number
                        shippingAddress:
                           type: object
                           properties:
                              street:
                                 type: string
                              city:
                                 type: string
                              postalCode:
                                 type: string
                        paymentMethod:
                           type: string
                           enum: [cash, card, stcpay]
                        discountCode:
                           type: string
         responses:
            "201":
               description: تم إنشاء الطلب بنجاح
            "400":
               description: خطأ في بيانات الطلب
...
   /orders/{orderId}:
      get:
         tags: [Orders]
         summary: تفاصيل طلب معين
         security:
            - bearerAuth: []
         parameters:
            - name: orderId
              in: path
              required: true
              schema:
                 type: string
         responses:
            "200":
               description: تفاصيل الطلب
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Order'
            "404":
               description: الطلب غير موجود
      delete:
         tags: [Orders]
         summary: إلغاء طلب معين
         security:
            - bearerAuth: []
         parameters:
            - name: orderId
              in: path
              required: true
              schema:
                 type: string
         responses:
            "200":
               description: تم إلغاء الطلب
            "404":
               description: الطلب غير موجود

   # Payment Routes
...
   /admin/dashboard/stats:
      get:
         tags: [Admin]
         summary: إحصائيات لوحة التحكم (للمدراء فقط)
         security:
            - bearerAuth: []
         responses:
            "200":
               description: إحصائيات النظام
               content:
                  application/json:
                     schema:
                        type: object
                        properties:
                           stats:
                              type: object
                              properties:
                                 totalOrders:
                                    type: integer
                                 totalUsers:
                                    type: integer
                                 totalProducts:
                                    type: integer
                                 pendingOrders:
                                    type: integer
                                 totalRevenue:
                                    type: number
            "403":
               description: غير مصرح لك بالوصول

   # Health Check
...
security:
   - bearerAuth: []
components:
   securitySchemes:
      bearerAuth:
         type: http
         scheme: bearer
         bearerFormat: JWT
   schemas:
...
      Product:
         type: object
         properties:
            _id:
               type: string
            name:
               type: string
            description:
               type: string
            price:
               type: number
            image:
               type: string
            category:
               type: string
            stock:
               type: integer
            isActive:
               type: boolean
            createdAt:
               type: string
               format: date-time
            updatedAt:
               type: string
               format: date-time
      ProductInput:
         type: object
         required: [name, price, category]
         properties:
            name:
               type: string
               example: "أسبرين 500 مجم"
            description:
               type: string
               example: "مسكن للألم ومضاد للالتهابات"
            price:
               type: number
               example: 25.50
            image:
               type: string
               example: "https://example.com/aspirin.jpg"
            category:
               type: string
               example: "60f1b2a3c4d5e6f7g8h9i0j1"
            stock:
               type: integer
               example: 100
            isActive:
               type: boolean
               default: true
      Category:
         type: object
         properties:
            _id:
               type: string
            name:
               type: string
            description:
               type: string
            image:
               type: string
            isActive:
               type: boolean
            createdAt:
               type: string
               format: date-time
            updatedAt:
               type: string
               format: date-time
      CategoryInput:
         type: object
         required: [name]
         properties:
            name:
               type: string
               example: "مسكنات الألم"
            description:
               type: string
               example: "أدوية لتسكين الألم والالتهابات"
            image:
               type: string
               example: "https://example.com/category.jpg"
            isActive:
               type: boolean
               default: true
...
      Order:
         type: object
         properties:
            _id:
               type: string
            user:
               type: string
            items:
               type: array
               items:
                  type: object
                  properties:
                     productId:
                        type: string
                     quantity:
                        type: integer
                     price:
                        type: number
                     product:
                        $ref: '#/components/schemas/Product'
            totalAmount:
               type: number
            shippingAddress:
               type: object
               properties:
                  street:
                     type: string
                  city:
                     type: string
                  postalCode:
                     type: string
            paymentMethod:
               type: string
               enum: [cash, card, stcpay]
            paymentStatus:
               type: string
               enum: [pending, paid, failed]
            orderStatus:
               type: string
               enum: [processing, shipped, delivered, cancelled]
            discountCode:
               type: string
            discountAmount:
               type: number
            createdAt:
               type: string
               format: date-time
            updatedAt:
               type: string
               format: date-time
...